fix: move generic method in another utils file
