+++ scripts/index.ts
@@ -1,6 +1,6 @@
 import dayjs from 'dayjs';
-import { readdirSync, readFileSync, writeFileSync } from 'fs';
 import * as path from 'path';
+import { readJSONFiles, writeJSONFiles } from '../lib/utils';
 
 import { AccountsEntity, FileEntity, Sample, TransactionsEntity } from '../types';
 
@@ -28,37 +28,6 @@ function mapAccount(nbOfDayToAdd: number) {
     ...account,
     balanceDate: dayjs(account.balanceDate).add(nbOfDayToAdd, 'day').toISOString(),
     transactions: account.transactions.map(mapTransactions(nbOfDayToAdd)),
-  });
-}
-
-/**
- * Read the samples directory and return an array of samples
- * @param dirPath Sample directory path
- */
-function readJSONFiles(dirPath: string): FileEntity[] {
-  const jsonsInDirectory: string[] = readdirSync(dirPath);
-  const samples: FileEntity[] = [];
-
-  jsonsInDirectory.forEach((filename: string) => {
-    const sample: Buffer = readFileSync(path.join(dirPath, filename));
-
-    samples.push({
-      filename,
-      sample: JSON.parse(sample.toString()),
-    });
-  });
-
-  return samples;
-}
-
-/**
- * Rewrite JSON samples with the refresh data
- * @param fileEntities Refresh sample file with file names
- * @param dirPath Samples directory path
- */
-function writeJSONFiles(fileEntities: FileEntity[], dirPath: string): void {
-  fileEntities.forEach((fileEntity: FileEntity) => {
-    writeFileSync(`${dirPath}/${fileEntity.filename}`, JSON.stringify(fileEntity.sample, null, 2));
   });
 }
